chat_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-decouple==3.8
httpx==0.25.2
openai==1.12.0
groq==0.4.1
tavily-python==0.3.0
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
easyocr==1.7.0
Pillow==10.1.0
scikit-learn==1.3.2
joblib==1.3.2
python-multipart==0.0.6
aiofiles==23.2.1

